+++ ydb/public/sdk/cpp/src/client/topic/impl/write_session_impl.h
@@ -443,9 +443,9 @@ private:
     void Connect(const TDuration& delay);
     void InitImpl();
     void ReadFromProcessor(); // Assumes that we're under lock.
-    void WriteToProcessorImpl(TClientMessage&& req); // Assumes that we're under lock.
+    void WriteToProcessorImpl(TClientMessage&& req, size_t requestMemoryUsage = 0); // Assumes that we're under lock.
     void OnReadDone(NYdbGrpc::TGrpcStatus&& grpcStatus, size_t connectionGeneration);
-    void OnWriteDone(NYdbGrpc::TGrpcStatus&& status, size_t connectionGeneration);
+    void OnWriteDone(NYdbGrpc::TGrpcStatus&& status, size_t connectionGeneration, size_t requestMemoryUsage);
     TProcessSrvMessageResult ProcessServerMessageImpl();
     TMemoryUsageChange OnMemoryUsageChangedImpl(i64 diff);
     TBuffer CompressBufferImpl(std::vector<std::string_view>& data, ECodec codec, i32 level);
@@ -519,6 +519,7 @@ private:
     std::string SessionId;
     IExecutor::TPtr CompressionExecutor;
     size_t MemoryUsage = 0; //!< Estimated amount of memory used
+    size_t WriteRequestsMemoryUsage = 0;
     bool FirstTokenSent = false;
 
     TMessageBatch CurrentBatch;